extract mutation for adding to cart and query for getting products
