+++ src/mutations/addToCart.js
@@ -1,0 +1,16 @@
+import gql from 'graphql-tag';
+import { graphql } from 'react-apollo';
+
+const query = gql`
+  mutation AddToCart($productId: Int!) {
+    addToCart(productId: $productId) @client
+  }
+`;
+
+const options = {
+  props: ({ mutate }) => ({
+    addToCart: product => mutate({ variables: { productId: product.id } }),
+  }),
+};
+
+export default graphql(query, options);
